ComplementarityModel

Documentation for ComplementarityModel.

ComplementarityModel.LinearComplementarityProblem

ComplementarityModel.MixedComplementarityProblem

ComplementarityModel.NonLinearComplementarityProblem

ComplementarityModel.CFB

ComplementarityModel.Cmin

ComplementarityModel.JCFB

ComplementarityModel.JCFBv

ComplementarityModel.JCmin

ComplementarityModel.JCminv

ComplementarityModel.LP_to_LCP

ComplementarityModel.LP_to_feasible_LCP

ComplementarityModel.LinearComplementarityProblem — Type

Structure to handle the linear complementarity problem in standard form:

0 ≤ Mx + q | x ≥ 0

ComplementarityModel.MixedComplementarityProblem — Type

Structure to handle the mixed complementarity problem in standard form:

0 ≤ F(x) | G(x) ≥ 0

ComplementarityModel.NonLinearComplementarityProblem — Type

Structure to handle the nonlinear complementarity problem in standard form:

0 ≤ F(x) | x ≥ 0

ComplementarityModel.CFB — Method

CFB(Fx, Gx)

Compute the C-function value.

ComplementarityModel.Cmin — Method

Cmin(Fx, Gx)

Compute the C-function value.

ComplementarityModel.JCFB — Method

JCFB(Fx, Gx, JFx, JGx)

Compute the jacobian's diagonal of the C-function.

ComplementarityModel.JCFBv — Method

JCFBv(Fx, Gx, JFx, JGx)

Compute the jacobian's diagonal of the C-function.

ComplementarityModel.JCmin — Method

JCmin(Fx, Gx, JFx, JGx)

Compute the jacobian's diagonal of the C-function.

ComplementarityModel.JCminv — Method

JCminv(Fx, Gx, JFx, JGx)

Compute the jacobian's diagonal of the C-function.

ComplementarityModel.LP_to_LCP — Method

LP_to_LCP(c, A, b)

Implements the self-embedding model for Linear Program

min cᵀx s.t.  Ax=b, x ≥ 0

from Roos, C., Terlaky, T., & Vial, J. P. (2005). Interior point methods for linear optimization.

ComplementarityModel.LP_to_feasible_LCP — Method

LP_to_feasible_LCP(c, A, b; μ = one(T))

Implements the self-embedding model for Linear Program

min cᵀx s.t.  Ax=b, x ≥ 0

from Roos, C., Terlaky, T., & Vial, J. P. (2005). Interior point methods for linear optimization. The returned problem is such that ones(nn) * √μ is on the central path.